Massive refactor in progress. BlinkOS compiles, next onto Blinklib
Blocking on IR data. What do you pass to userland? Full buffers? I think so.
Then how do we intercept out-of-band stuff for the OS layer? Stuff like new game load, and all-off? Add an out-of-band marker somewhere? But where?
