+++ AS7/blink/blinkstate/blinklib.componentinfo.xml
