+++ AS7/blinkos/blinkos/blinkos.cppproj
@@ -4,7 +4,7 @@
     <SchemaVersion>2.0</SchemaVersion>
     <ProjectVersion>7.0</ProjectVersion>
     <ToolchainName>com.Atmel.AVRGCC8.CPP</ToolchainName>
-    <ProjectGuid>{c7103c5c-377a-4589-89b5-ee449b21a39a}</ProjectGuid>
+    <ProjectGuid>{75c30a81-edd6-4de5-8d7e-d379aa89f81a}</ProjectGuid>
     <avrdevice>ATmega168PB</avrdevice>
     <avrdeviceseries>none</avrdeviceseries>
     <OutputType>StaticLibrary</OutputType>
@@ -14,9 +14,9 @@
     <OutputDirectory>$(MSBuildProjectDirectory)\$(Configuration)</OutputDirectory>
     <AvrGccProjectExtensions>
     </AvrGccProjectExtensions>
-    <AssemblyName>blinkstate</AssemblyName>
-    <Name>blinkstate</Name>
-    <RootNamespace>blinkstate</RootNamespace>
+    <AssemblyName>blinkos</AssemblyName>
+    <Name>blinkos</Name>
+    <RootNamespace>blinkos</RootNamespace>
     <ToolchainFlavour>Native</ToolchainFlavour>
     <KeepTimersRunning>true</KeepTimersRunning>
     <OverrideVtor>false</OverrideVtor>
@@ -30,21 +30,17 @@
     <ResetRule>0</ResetRule>
     <eraseonlaunchrule>0</eraseonlaunchrule>
     <EraseKey />
-    <AAFDebugger>
-      <AAFDebugFiles xmlns="">
-      </AAFDebugFiles>
-    </AAFDebugger>
     <AsfFrameworkConfig>
       <framework-data xmlns="">
-        <options />
-        <configurations />
-        <files />
-        <documentation help="" />
-        <offline-documentation help="" />
-        <dependencies>
-          <content-extension eid="atmel.asf" uuidref="Atmel.ASF" version="3.35.1" />
-        </dependencies>
-      </framework-data>
+  <options />
+  <configurations />
+  <files />
+  <documentation help="" />
+  <offline-documentation help="" />
+  <dependencies>
+    <content-extension eid="atmel.asf" uuidref="Atmel.ASF" version="3.40.0" />
+  </dependencies>
+</framework-data>
     </AsfFrameworkConfig>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
@@ -119,7 +115,6 @@
   </avrgcc.compiler.symbols.DefSymbols>
   <avrgcc.compiler.directories.IncludePaths>
     <ListValues>
-      <Value>../../../../variants/standard</Value>
       <Value>%24(PackRepoDir)\atmel\ATmega_DFP\1.2.209\include</Value>
     </ListValues>
   </avrgcc.compiler.directories.IncludePaths>
@@ -137,9 +132,9 @@
   </avrgcccpp.compiler.symbols.DefSymbols>
   <avrgcccpp.compiler.directories.IncludePaths>
     <ListValues>
-      <Value>../../../../libraries/blinklib/src</Value>
-      <Value>../../../../cores/blinkcore</Value>
       <Value>%24(PackRepoDir)\atmel\ATmega_DFP\1.2.209\include</Value>
+      <Value>../../../../cores/blinkcore</Value>
+      <Value>../../../../variants/standard</Value>
     </ListValues>
   </avrgcccpp.compiler.directories.IncludePaths>
   <avrgcccpp.compiler.optimization.level>Optimize (-O1)</avrgcccpp.compiler.optimization.level>
@@ -162,14 +157,45 @@
     </ToolchainSettings>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="..\..\..\libraries\blinkstate\src\blinkstate.cpp">
+    <Compile Include="..\..\..\libraries\blinkos\src\blinkos.cpp">
       <SubType>compile</SubType>
-      <Link>blinkstate.cpp</Link>
+      <Link>blinkos.cpp</Link>
     </Compile>
-    <Compile Include="..\..\..\libraries\blinkstate\src\blinkstate.h">
+    <Compile Include="..\..\..\libraries\blinkos\src\blinkos.h">
       <SubType>compile</SubType>
-      <Link>blinkstate.h</Link>
+      <Link>blinkos.h</Link>
     </Compile>
+    <Compile Include="..\..\..\libraries\blinkos\src\blinkos_button.cpp">
+      <SubType>compile</SubType>
+      <Link>blinkos_button.cpp</Link>
+    </Compile>
+    <Compile Include="..\..\..\libraries\blinkos\src\blinkos_button.h">
+      <SubType>compile</SubType>
+      <Link>blinkos_button.h</Link>
+    </Compile>
+    <Compile Include="..\..\..\libraries\blinkos\src\blinkos_irdata.cpp">
+      <SubType>compile</SubType>
+      <Link>blinkos_irdata.cpp</Link>
+    </Compile>
+    <Compile Include="..\..\..\libraries\blinkos\src\blinkos_irdata.h">
+      <SubType>compile</SubType>
+      <Link>blinkos_irdata.h</Link>
+    </Compile>
+    <Compile Include="..\..\..\libraries\blinkos\src\blinkos_timer.cpp">
+      <SubType>compile</SubType>
+      <Link>blinkos_timer.cpp</Link>
+    </Compile>
+    <Compile Include="..\..\..\libraries\blinkos\src\blinkos_timer.h">
+      <SubType>compile</SubType>
+      <Link>blinkos_timer.h</Link>
+    </Compile>
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\blinkcore\blinkcore.cppproj">
+      <Name>blinkcore</Name>
+      <Project>{dce6c7e3-ee26-4d79-826b-08594b9ad897}</Project>
+      <Private>True</Private>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(AVRSTUDIO_EXE_PATH)\\Vs\\Compiler.targets" />
 </Project>
