--- cores/blinkcore/run.h
@@ -1,16 +1,0 @@
-/*
- * run.h
- *
- * Prototypes the run() function that is called after all hardware
- * has been initialized and the blinkcore functions are operational.
- */
-
-
-#ifndef RUN_H_
-#define RUN_H_
-
-// Called repeatedly after hardware has been initialized.
-
-void run(void);
-
-#endif
