+++ libraries/blinklib/library.properties
@@ -1,8 +1,8 @@
-name=Blink Developement Interface
+name=Blink Library
 version=1.0
 author=Move38.com
 maintainer=Move38.com
-sentence=High level functions for blinks game developement
+sentence=Code that gets linked into games to talk to the OS and provide utility functions
 paragraph=
 category=Other
 url=https://github.com/bigjosh/Move38-Arduino-Platform
