+++ libraries/blinklib/src/Arduino.h
@@ -88,8 +88,6 @@
 	// manual #include in their sketches would be to throw all the library code into one giant
 	// directory... and that is even uglier, right?
 
-	#include "shared.h"
 	#include "blinklib.h"
-	#include "blinkstate.h"
 
 #endif
