+++ libraries/blinklib/src/Serial.h
@@ -16,7 +16,7 @@
 
     #include "Print.h"
 
-    #define DEF_SERVICE_PORT_BAUDRATE 500000
+    #define DEF_SERVICE_PORT_BAUDRATE 1000000
 
     class ServicePortSerial : public Print
     {
