+++ libraries/blinkos/library.properties
@@ -1,8 +1,8 @@
-name=Blink Stateful Interface
+name=Blink Developement Interface
 version=1.0
 author=Move38.com
 maintainer=Move38.com
-sentence=High level state model view for blinks game developement
+sentence=High level interface between games and the hardware core
 paragraph=
 category=Other
 url=https://github.com/bigjosh/Move38-Arduino-Platform
