+++ libraries/blinkos/src/test.cpp
@@ -1,0 +1,14 @@
+#include "blinkos.h"
+
+void setupEntry() {
+    
+    // Blank
+    
+}
+
+
+void loopEntry( loopstate_in_t const *loopstate_in , loopstate_out_t *loopstateout ){
+    
+    
+}    
+
